move to writer
